refactor react components
