+++ dialogs_front/containers/App.js
@@ -1,0 +1,8 @@
+import { connect } from "react-redux";
+import App from "../components/App";
+
+const mapStateToProps = state => ({
+  status: state.status
+});
+
+export default connect(mapStateToProps)(App);
