+++ dialogs_front/containers/DialogList.js
@@ -1,0 +1,8 @@
+import { connect } from "react-redux";
+import * as actions from "../actions/viewActions";
+import DialogList from "../components/DialogList";
+
+export default connect(
+  () => ({}),
+  actions
+)(DialogList);
